fix: cancel stream listeners on drop and stop reporting timeouts as end of stream
Each stream box started its listener with `Task { [weak self] in guard let
self ... }`, which holds the box strongly for the whole `for await` loop.
The box's deinit, the only place that cancelled the task, could therefore
never run: dropping Transaction::updates() and the storefront, purchase
intent and subscription status streams leaked the box, kept the listener
running and let its queue grow without bound. All five streams now share
one generic SKStreamBox. Its listener task captures only the queue, so
dropping the Rust handle deinitializes the box, which cancels the task and
closes the queue.

A timeout was also reported as Ok(None), the same value as the end of the
sequence, so `while let Some(t) = updates.next()?` stopped listening after
30 s of quiet and missed renewals, refunds and Ask to Buy approvals.
next() now waits until the next value arrives or the sequence ends, and
next_timeout() returns StoreKitError::TimedOut when no value arrives in
time; the stream stays open. The deadline is fixed for the whole wait
instead of restarting after every wakeup. Filtered transaction streams
also reject a missing product ID instead of silently returning every
transaction.

Copilot-Session: 61abddcf-4f9d-46ad-ba0d-7a0382752af1

+++ src/ffi.rs
@@ -45,7 +45,7 @@ extern "C" {
     pub fn sk_transaction_stream_release(stream: *mut c_void);
     pub fn sk_transaction_stream_next(
         stream: *mut c_void,
-        timeout_ms: u32,
+        timeout_ms: i64,
         out_transaction: *mut *mut c_void,
         out_verification_json: *mut *mut c_char,
         out_error_message: *mut *mut c_char,
@@ -108,7 +108,7 @@ extern "C" {
     pub fn sk_storefront_stream_release(stream: *mut c_void);
     pub fn sk_storefront_stream_next(
         stream: *mut c_void,
-        timeout_ms: u32,
+        timeout_ms: i64,
         out_storefront_json: *mut *mut c_char,
         out_error_message: *mut *mut c_char,
     ) -> i32;
@@ -134,7 +134,7 @@ extern "C" {
     pub fn sk_subscription_status_stream_release(stream: *mut c_void);
     pub fn sk_subscription_status_stream_next(
         stream: *mut c_void,
-        timeout_ms: u32,
+        timeout_ms: i64,
         out_status_json: *mut *mut c_char,
         out_error_message: *mut *mut c_char,
     ) -> i32;
@@ -144,7 +144,7 @@ extern "C" {
     pub fn sk_subscription_group_status_stream_release(stream: *mut c_void);
     pub fn sk_subscription_group_status_stream_next(
         stream: *mut c_void,
-        timeout_ms: u32,
+        timeout_ms: i64,
         out_payload_json: *mut *mut c_char,
         out_error_message: *mut *mut c_char,
     ) -> i32;
@@ -164,7 +164,7 @@ extern "C" {
     pub fn sk_purchase_intent_stream_release(stream: *mut c_void);
     pub fn sk_purchase_intent_stream_next(
         stream: *mut c_void,
-        timeout_ms: u32,
+        timeout_ms: i64,
         out_payload_json: *mut *mut c_char,
         out_error_message: *mut *mut c_char,
     ) -> i32;
